requirements.txt
Flask==3.1.0
flask-cors==5.0.1
google-genai==1.13.0
python-dotenv==1.0.1
